+++ multiqc/modules/busco/busco.py
@@ -5,7 +5,7 @@ import logging
 import re
 from collections import OrderedDict
 
-from multiqc.modules.base_module import BaseMultiqcModule
+from multiqc.modules.base_module import BaseMultiqcModule, ModuleNoSamplesFound
 from multiqc.plots import bargraph
 
 # Initialise the logger
@@ -46,7 +46,7 @@ class MultiqcModule(BaseMultiqcModule):
         self.busco_data = self.ignore_samples(self.busco_data)
 
         if len(self.busco_data) == 0:
-            raise UserWarning
+            raise ModuleNoSamplesFound
 
         log.info("Found {} reports".format(len(self.busco_data)))