+++ multiqc/modules/gatk/varianteval.py
@@ -24,6 +24,10 @@ class VariantEvalMixin:
                 self.add_data_source(f, section="varianteval")
                 self.gatk_varianteval[f["s_name"]] = parsed_data
 
+            # Superfluous function call to confirm that it is used in this module
+            # Replace None with actual version if it is available
+            self.add_software_version(None, f["s_name"])
+
         # Filter to strip out ignored sample names
         self.gatk_varianteval = self.ignore_samples(self.gatk_varianteval)