+++ multiqc/modules/macs2/macs2.py
@@ -5,7 +5,7 @@ import logging
 import re
 from collections import OrderedDict
 
-from multiqc.modules.base_module import BaseMultiqcModule
+from multiqc.modules.base_module import BaseMultiqcModule, ModuleNoSamplesFound
 from multiqc.plots import bargraph
 
 # Initialise the logger
@@ -33,7 +33,7 @@ class MultiqcModule(BaseMultiqcModule):
         self.macs_data = self.ignore_samples(self.macs_data)
 
         if len(self.macs_data) == 0:
-            raise UserWarning
+            raise ModuleNoSamplesFound
 
         log.info("Found {} logs".format(len(self.macs_data)))
         self.write_data_file(self.macs_data, "multiqc_macs")