+++ multiqc/modules/methylQA/methylQA.py
@@ -6,7 +6,7 @@ import os
 import re
 from collections import OrderedDict
 
-from multiqc.modules.base_module import BaseMultiqcModule
+from multiqc.modules.base_module import BaseMultiqcModule, ModuleNoSamplesFound
 from multiqc.plots import linegraph
 
 # Initialise the logger
@@ -32,11 +32,15 @@ class MultiqcModule(BaseMultiqcModule):
         for f in self.find_log_files("methylQA"):
             self.parse_methylqa_logs(f)
 
+            # Superfluous function call to confirm that it is used in this module
+            # Replace None with actual version if it is available
+            self.add_software_version(None, f["s_name"])
+
         # Filter to strip out ignored sample names
         self.methylqa_data = self.ignore_samples(self.methylqa_data)
 
         if len(self.methylqa_data) == 0:
-            raise UserWarning
+            raise ModuleNoSamplesFound
 
         log.info("Found {} reports".format(len(self.methylqa_data)))