+++ multiqc/modules/ngsderive/ngsderive.py
@@ -6,7 +6,7 @@ import io
 import logging
 from collections import OrderedDict
 
-from multiqc.modules.base_module import BaseMultiqcModule
+from multiqc.modules.base_module import BaseMultiqcModule, ModuleNoSamplesFound
 from multiqc.plots import bargraph, linegraph, table
 
 # Initialise the logger
@@ -48,6 +48,10 @@ class MultiqcModule(BaseMultiqcModule):
                 "strandedness",
                 expected_header_count_strandedness,
             )
+
+            # Superfluous function call to confirm that it is used in this module
+            # Replace None with actual version if it is available
+            self.add_software_version(None, f["s_name"])
 
         for f in self.find_log_files("ngsderive/instrument"):
             self.parse(
@@ -91,7 +95,7 @@ class MultiqcModule(BaseMultiqcModule):
             [len(d) for d in [self.strandedness, self.instrument, self.readlen, self.encoding, self.junctions]]
         )
         if num_results_found == 0:
-            raise UserWarning
+            raise ModuleNoSamplesFound
         log.info("Found {} reports".format(num_results_found))
 
         if self.strandedness: