+++ multiqc/modules/rseqc/junction_saturation.py
@@ -41,6 +41,10 @@ def parse_reports(self):
                     self.junction_saturation_known[f["s_name"]][v] = parsed["y"][k]
                     self.junction_saturation_novel[f["s_name"]][v] = parsed["w"][k]
 
+        # Superfluous function call to confirm that it is used in this module
+        # Replace None with actual version if it is available
+        self.add_software_version(None, f["s_name"])
+
     # Filter to strip out ignored sample names
     self.junction_saturation_all = self.ignore_samples(self.junction_saturation_all)
     self.junction_saturation_known = self.ignore_samples(self.junction_saturation_known)